Move utility functions to designated submodule readers._utils._image
